#184 - AccountingHttp shows now real value of quota from API (#527)
* #184 - AccountingHttp shows now real value of quota from API

* #184 - changed quota string into long

* extracted interface for accounting http

* stub class created

* wrong exception text

* Renamed to BoxClient and BoxHttp

* added size value into quotaDto, the client returns now the Object instead of just a long value.
changed in BoxVolumeBlock declarations of the interface instead of concrete class

* EMAIL_KEY doesnt need to be in interface swiched back to concrete

* renamed QuotaDto into QuotaState

* refactoring did too much in here

* interface cleaned

* interface reverted

* added calculations of free quota and quota

* fixes for desktop integration

* clean up

* clean up

* rename

* added quotaDescription test

* navigate opened

* interface for BoxVolume extracted, increased version in build.gradle

* added downwards compatibility constructor

* added config in interface

* QuotaState properties private, code cleanup

+++ core/src/main/java/de/qabel/core/accounting/AccountingProfile.java
@@ -5,24 +5,24 @@ import de.qabel.core.config.SyncSettingItem;
 import java.util.ArrayList;
 
 public class AccountingProfile extends SyncSettingItem {
-    private Integer quota;
+    private long quota;
     private ArrayList<String> prefixes;
 
     public AccountingProfile() {
         prefixes = new ArrayList<>();
     }
 
-    public AccountingProfile(int quota, String prefix) {
+    public AccountingProfile(long quota, String prefix) {
         this.quota = quota;
         prefixes = new ArrayList<>();
         prefixes.add(prefix);
     }
 
-    public Integer getQuota() {
+    public long getQuota() {
         return quota;
     }
 
-    public void setQuota(int quota) {
+    public void setQuota(long quota) {
         this.quota = quota;
     }
